# dist/openapi.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: Localhost (dev)
    url: http://localhost:3000
info:
  description: This is an API designed for users to create, share and explore recipes and create family recipes
  version: 1.0.2
  title: Recipe Sharing API
  contact:
    name: Abed
    email: showganabed@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /users/familyRecipes:
    get:
      tags: [family]
      summary: Get family recipes for a specific user

      responses:
        '200':
          description: A list of family recipes for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FamilyRecipe'

    post:
      tags: [family]
      summary: Add a new family recipe for a user

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyRecipe'
      responses:
        '201':
          description: Family recipe created
...
  /users/favoriteRecipes/{recipeId}:         
    post:
      tags:
        - recipes
      summary: Add a recipe to user's favorites
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
          description: ID of the recipe to add
      responses:
        '200':
          description: Recipe marked as favorite successfully
        '401':
          description: User not logged in
        '404':
          description: Recipe not found locally or in Spoonacular
        '500':
          description: Server error
      security:
        - sessionAuth: []

    delete:
      tags: [recipes]
      summary: Remove a recipe from user's favorites
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
          description: ID of the recipe to delete
      responses:
        '200':
          description: Recipe removed from favorites
  /users/last:
    get:
      summary: Get last 3 viewed recipe IDs
      tags:
        - Recipes
      responses:
        '200':
          description: A list of up to 3 last viewed recipe IDs (most recent first)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example: ["abc123", "def456", "ghi789"]
        '401':
          description: User not logged in
        '500':
          description: Server error while fetching recipes
  /users/clickOnRecipe:
    post:
      summary: Add last clicked recipe (max 3)
      tags:
        - Recipes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipeId
              properties:
                recipeId:
                  type: string
                  example: "12345"
      responses:
        '200':
          description: Click saved successfully
        '500':
          description: Server error while saving click
...
  /recipes/{recipeId}:
    get:
      tags:
        - recipes
      summary: Get full recipe details
      operationId: getRecipeById
      description: |
        Returns the full recipe details including preview fields, ingredients list,
        instructions, and number of meals.
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
          description: ID of the recipe to retrieve
      responses:
        '200':
          description: Full recipe details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          description: Recipe not found

    delete:
      tags:
        - recipes
      summary: Delete a recipe by ID
      operationId: deleteRecipe
      parameters:
        - name: recipeId
          in: path
          required: true
          description: ID of the recipe to delete from the database
          schema:
            type: integer
      responses:
        '200':
          description: Recipe deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Recipe not found or unauthorized
components:
  securitySchemes: 
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT
  schemas:
...
    NewRecipe:
      type: object
      required:
        - recipe_image
        - recipe_title
        - prep_duration
        - popularity
        - vegetarian
        - vegan
        - gluten_free
        - ingredients
        - instructions
        - amount_of_meals
      properties:
        recipe_image:
          type: string
          format: uri
          example: https://example.com/image.jpg
        recipe_title:
          type: string
          example: Chocolate Cake
        prep_duration:
          type: string
          example: 00:45:00
          description: Preparation time in HH:MM:SS format
        popularity:
          type: integer
          example: 152
        vegetarian:
          type: boolean
        vegan:
          type: boolean
        gluten_free:
          type: boolean
        clicked_by_user:
          type: boolean
        saved_by_user:
          type: boolean
        ingredients:
          type: array
          description: List of ingredients and their amounts
          items:
            type: object
            required: [name, amount]
            properties:
              name:
                type: string
                example: Sugar
              amount:
                type: string
                example: "2 cups"
        instructions:
          type: string
          example: |
            1. Preheat the oven to 180°C.
            2. Mix all ingredients.
            3. Bake for 30 minutes.
        amount_of_meals:
          type: integer
          example: 4
    Recipe:
      allOf:
        - $ref: '#/components/schemas/NewRecipe'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              example: 123e4567-e89b-12d3-a456-426614174000
    FamilyRecipe:
      type: object
      required: [recipe_author, recipe_season, recipe_ingredients, recipe_instructions]
      properties:
        recipe_author:
          type: string
        recipe_season:
          type: string
        recipe_ingredients:
          type: array
          items:
            type: string
        recipe_instructions:
          type: string
...